+++ packages/dunya/test/country_model_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:dunya/dunya.dart';
 
 void main() {