+++ packages/dunya/test/country_search_test.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_test/flutter_test.dart';
+import 'package:test/test.dart';
 import 'package:dunya/src/models/country.dart';
 import 'package:dunya/src/search/country_search.dart';